Added some basic but often used supersource layouts
This includes a PiP for powerpoint with a camera on box 1 and 2 of supersource 1.

And:
A side by side for a camera with remote(or anything else)

These are controlled using macro's
macro 0 enables box 1 and 2, disables 3 and 4 of SS1
macro 5 enables box 3 and 4, disables 1 and 2 of SS1

the other macro's control the box source of the camera, switch them between cam 1,2,3

Also included the SME background and some alignment grids.